lyrics

+++ app/src/main/java/com/example/music_room/data/remote/model/PlaybackStateDto.kt
@@ -4,7 +4,7 @@ import com.squareup.moshi.Json
 
  data class PlaybackStateDto(
     @Json(name = "currentTrack") val currentTrack: TrackDto?,
-    @Json(name = "positionSeconds") val positionSeconds: Int,
+     @Json(name = "positionSeconds") val positionSeconds: Double,
     @Json(name = "isPlaying") val isPlaying: Boolean,
     @Json(name = "queue") val queue: List<TrackDto>,
     @Json(name = "shuffleEnabled") val shuffleEnabled: Boolean,
